python3.12/django/drf/app_02/pyproject.toml
[tool.poetry]
name = "app-02"
version = "0.1.0"
description = ""
authors = ["noskill <golanghack@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
